feat: AddEventView, EventDetailView 구현
