+++ Muckmuck/MuckmuckApp.swift
@@ -6,12 +6,14 @@
 //
 
 import SwiftUI
+import SwiftData
 
 @main
 struct MuckmuckApp: App {
     var body: some Scene {
         WindowGroup {
-            ContentView()
+            HomeView()
         }
+        .modelContainer(ModelContainerProvider.shared)
     }
 }
