+++ Muckmuck/View/AddEventView.swift
@@ -21,43 +21,72 @@ struct AddEventView: View {
                 Image(.beer)
                     .resizable()
                     .frame(width: 146, height: 228)
+                    .padding()
                 
                 ZStack {
                     RoundedRectangle(cornerRadius: 30)
-                        .frame(width: UIScreen.mainBounds.width, height: 200)
+                        .frame(width: UIScreen.mainBounds.width-32, height: 200)
                         .foregroundStyle(.white.opacity(0.5))
                     VStack(spacing: 16) {
-                        TextField("이벤트 제목", text: $eventName)
+                        TextField("모임 제목", text: $eventName)
                             .background(.clear)
                             .cornerRadius(12)
+                            .multilineTextAlignment(.center)
+                            .font(.title)
+                        
                         Divider()
-                        DatePicker("날짜 및 시간", selection: $date, displayedComponents: [.date, .hourAndMinute])
-                            .datePickerStyle(.compact)
+                        
+                        Button(action: {
+                            showDatePicker.toggle()
+                        }, label: {
+                            HStack {
+                                Image(systemName: "calendar.badge.plus")
+                                Text("날짜 및 시간")
+                                    .font(.title2)
+                            }
+                        })
+                        
                         Divider()
                         Button {
-                            // showLocationSearch.toggle()
                         } label: {
                             HStack {
                                 Image(systemName: "mappin.and.ellipse")
                                 Text("위치")
+                                    .font(.title2)
                             }
                         }
                     }
                 }
+                .padding([.horizontal, .bottom])
+                
                 
                 ZStack {
                     RoundedRectangle(cornerRadius: 30)
-                        .frame(width: UIScreen.mainBounds.width, height: 140)
+                        .frame(width: UIScreen.mainBounds.width - 32, height: 140)
                         .foregroundStyle(.white.opacity(0.5))
                     VStack(spacing: 16) {
                         Stepper(value: $participantCount, in: 1...20) {
-                            Text("최소 인원 수: \(participantCount)명")
-                        }
+                            Text("최소 인원: \(participantCount)명")
+                                .font(.title)
+                        }.frame(width:UIScreen.mainBounds.width - 100)
                         Divider()
                         Stepper(value: $participantCount, in: 1...20) {
-                            Text("최대 인원 수: \(participantCount)명")
+                            Text("최대 인원: \(participantCount)명")
+                                .font(.title)
                         }
+                        .frame(width:UIScreen.mainBounds.width - 100)
                     }
+                }
+                .padding([.horizontal, .bottom])
+                .sheet(isPresented: $showDatePicker) {
+                    VStack {
+                        DatePicker("날짜 및 시간", selection: $date, displayedComponents: [.date, .hourAndMinute]
+                        )
+                        .datePickerStyle(.compact)
+                        .padding()
+                        
+                    }
+                    .presentationDetents([.height(100)])
                 }
                 
                 Button(action: {
